🔀: 내 근로 정리(WorkArrangement)페이지 css를 마무리한다.  (#34)
* 🔧 : propTypes를 사용하기 위한 타입 지정 라이브러리 설치

* 🔧: IDE가 아닌 프로젝트 vsc settings.json 설정

* 🧪 : contextAPI를 활용해 전역 상태 변수 커스텀 테스트, 각 페이지에서 context 전역 상태 변수를 사용하지 않고 가독성을 위한 커스텀 훅 제작

* ✨: 제작한 context provider 컴포넌트 최상위 커스텀에 설정

* ✨: 커스텀 훅을 사용하지 않고 context Provider 컴포넌트만 생성

* ✨: context provider로 부터 생성된 여러 상태변수를 DevTerminal에 적용 예시, useContext 사용

* 🐛: prop-types는 dev가 아닌 dependencies에서 사용된다. 개발을 위한 것이 아니다.

* ✨: 코드 가독성을 위해 useContext 등을 제외한 useGlobalState 커스텀 훅을 제작하고 사용

* 💄: Main 페이지 Scss 적용 및 확인 완료

* 💄: 내 근로 정리 WorkArrangement 페이지 css를 마무리한다.

* 💄: 하단 버튼 위치 변경

* ♻️: 팀원과 협의된 디렉토리 구조 적용
