+++ src/Page/WorkArrangement/components/WorkArrangement.jsx
@@ -1,33 +1,36 @@
 import React from 'react';
-import quizImage from '../image/quiz-short-cut.png';
-import arrowRight from '../image/arrow-right.png';
-import '../css/WorkArrangement.css';
+import quizImage from '../image/quiz-short-cut.svg';
+import arrowRight from '../image/arrow-right.svg';
+import '../css/WorkArrangement.scss';
 
 const WorkArrangement = () => {
   return (
     <div className="work-arrangement">
-      <div className="container">
-        <h1>현재 나의 근로 상황 얼마나 알고 계시나요?</h1>
-        <div className="quiz">
-          <img src={quizImage} alt="quiz-name" />
-          <div className="short-cut">바로가기</div>
+      <div className="wa-container">
+        <div id="wa-title">
+          <div>
+            현재 <span>나의 근로 상황</span>
+          </div>
+          <div>알고 계시나요?</div>
         </div>
-        <div className="contents">
-          <h1>다른 결과지들은 어떨까요?</h1>
-          <div className="result">
-            <div className="control">
-              <h3>미도인 성수 근로 결과지</h3>
-              <div id="short-cut">
+        <div className="wa-quiz">
+          <img src={quizImage} alt="quiz-name" />
+          <div id="wa-short-cut">바로가기</div>
+        </div>
+        <div className="wa-contents">
+          <div className="wa-title">다른 결과지들은 어떨까요?</div>
+          <div className="wa-result">
+            <div className="wa-control">
+              <div className="wa-title">미도인 성수 근로 결과지</div>
+              <div id="wa-short-cut">
                 <a href="/#">상세보기</a>
-                <img src={arrowRight} alt="short-cut" />
+                <div>
+                  <img src={arrowRight} alt="short-cut" />
+                </div>
               </div>
             </div>
-            <div className="group">
-              <div className="description">
-                <div>
-                  상시 5인 미만 사업장에서 근무하시므로 추가적인 가산 수당이
-                  없습니다.
-                </div>
+            <div className="wa-group">
+              <div className="wa-description">
                 <div>
                   상시 5인 미만 사업장에서 근무하시므로 추가적인 가산 수당이
                   없습니다.
@@ -45,15 +48,19 @@ const WorkArrangement = () => {
                   없습니다.
                 </div>
               </div>
-              <div className="options">
-                <div>발생 요건들</div>
-                <div id="option">주휴수당</div>
+              <div className="wa-options">
+                <div className="wa-contents">
+                  <div className="wa-title">발생 요건들</div>
+                  <div id="wa-option">주휴수당</div>
+                </div>
               </div>
             </div>
           </div>
-          <a id="short-cut" href="/#">
-            더보기
-          </a>
+          <div className="wa-group">
+            <div className="wa-short-cut" href="/#">
+              더보기
+            </div>
+          </div>
         </div>
       </div>
     </div>
