+++ src/Page/WorkArrangement/css/WorkArrangement.scss
@@ -1,0 +1,195 @@
+.work-arrangement {
+  background-color: #EFF6FE;
+  display: flex;
+  box-sizing: border-box;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  min-height: 100vh;
+  font-family: 'Pretendard';
+
+  .wa-container {
+    /* background-color: blueviolet; */
+    width: 1180px;
+
+    #wa-title {
+      margin: 90px 0 60px 0;
+      font-size: 48px;
+      font-weight: 600;
+      line-height: 68px;
+
+      span {
+        color: #1BBFC1;
+      }
+    }
+
+    .wa-quiz {
+      /* background-color: bisque; */
+      display: flex;
+      box-sizing: border-box;
+      flex-direction: row;
+      margin-bottom: 140px;
+
+      #wa-short-cut {
+        background-color: #204598;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        color: white;
+        width: 280px;
+        height: 308px;
+        margin-left: 20px;
+        font-weight: 600;
+        font-size: 32px;
+        line-height: 42px;
+      }
+    }
+
+    // #wa-short-cut {
+    //   background-color: #1BBFC1;
+    //   display: flex;
+    //   justify-content: center;
+    //   align-items: center;
+    //   width: 73px;
+    //   height: 56px;
+    //   margin-left: 93.81%;
+    // }
+    .wa-contents {
+      .wa-title {
+        font-weight: 600;
+        font-size: 48px;
+        line-height: 63px;
+        margin-bottom: 60px;
+      }
+
+      /* background-color: brown; */
+      .wa-result {
+        background-color: white;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        margin-bottom: 40px;
+        box-shadow: 0px 4px 10px -1px rgba(35, 48, 56, 0.08);
+        padding: 40px;
+
+        .wa-control {
+          // background-color: aquamarine;
+          width: 100%;
+          display: flex;
+          justify-content: space-between;
+          align-items: center;
+          margin-bottom: 8px;
+
+          .wa-title {
+            font-size: 32px;
+            font-weight: 600;
+            line-height: 42px;
+            margin: 0;
+          }
+
+          #wa-short-cut {
+            // background-color: yellow;
+            display: flex;
+            flex-direction: row;
+            font-size: 23.04px;
+            font-weight: 400;
+            line-height: 32.92px;
+            width: 98px;
+            height: 33px;
+
+            a {
+              text-decoration-line: none;
+              color: #111;
+              font-size: 23.041px;
+            }
+          }
+
+          h3 {
+            margin: 0;
+            font-size: 32px;
+          }
+        }
+
+        .wa-group {
+          // background-color: antiquewhite;
+          width: 100%;
+          display: flex;
+          justify-content: space-between;
+          flex-direction: row;
+          align-items: center;
+
+          .wa-description {
+            display: flex;
+            flex-direction: column;
+
+            /* background-color: yellowgreen; */
+            // margin: 0 111px 46px 40px;
+            div {
+              line-height: 38px;
+            }
+          }
+
+          .wa-options {
+            display: flex;
+            flex-direction: column;
+            width: 450px;
+
+            /* background-color: tomato; */
+            /* border: 1px solid black; */
+            // margin-left: 111px;
+            .wa-contents {
+              padding-left: 111px;
+              border-left: 2px solid #E5E5EC;
+              height: 158px;
+
+              .wa-title {
+                font-size: 18px;
+                font-weight: 600;
+                line-height: 28px;
+                margin-top: 14px;
+                margin-bottom: 12px;
+              }
+
+              #wa-option {
+                background-color: #204598;
+                font-size: 12px;
+                color: white;
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                border-radius: 50px;
+                width: 95px;
+                height: 45px;
+                margin-right: 12px;
+                margin-bottom: 24px;
+              }
+
+            }
+          }
+        }
+      }
+
+      .wa-group {
+        width: 100%;
+        display: flex;
+        justify-content: flex-end;
+
+        .wa-short-cut {
+          background-color: #1BBFC1;
+          display: flex;
+          justify-content: center;
+          align-items: center;
+          color: #FFFFFF;
+          width: 73px;
+          height: 56px;
+          margin-top: 20px;
+          font-weight: 600;
+          font-size: 16px;
+          line-height: 24px;
+        }
+      }
+    }
+  }
+}
